requirements.txt
# Core LangChain packages
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0

# Web framework (Streamlit only)
streamlit>=1.28.0

# Environment and configuration
python-dotenv>=1.0.0

# Vector database
faiss-cpu>=1.8.0

# OpenAI utilities
tiktoken>=0.5.0

# Compatibility
protobuf>=4.0.0

# Data processing
pandas>=2.0.0

# Optional: Plotting and visualization (for enhanced analytics dashboard)
# plotly>=5.15.0